Rename wmt.py to translate.py split language pairs for clarity dissociate ende / enfr, make them independant
