+++ tensor2tensor/bin/t2t-datagen
