+++ tensor2tensor/bin/t2t-decoder
